+++ src/Application/Query/Happening/Webinar/GetWebinarViewQueryHandler.php
@@ -59,6 +59,7 @@ class GetWebinarViewQueryHandler
             round($timeRemainingInSeconds * 0.2),
             $happening->getWebinarHeaderImage($query->getLocale()),
             $happening->getLiveUrl(),
+            $happening->isSidebarAllowed(),
             $this->isVideoWebinarAndHappeningIsEnded($happening)
         );
     }