+++ tests/Ui/Bundle/AdminBundle/Controller/Happening/UpdateActionTest.php
@@ -164,6 +164,7 @@ class UpdateActionTest extends TestCase
         $this->happening->isInteractiveWebinar()->shouldBeCalled()->willReturn(false);
         $this->happening->isVideoWebinar()->shouldBeCalled()->willReturn(false);
         $this->happening->getLiveUrl()->shouldBeCalled()->willReturn(null);
+        $this->happening->isSidebarAllowed()->willReturn(true);
 
         $form = $this->prophesize(Form::class);
         $formView = $this->prophesize(FormView::class);
@@ -238,6 +239,7 @@ class UpdateActionTest extends TestCase
         $this->happening->isInteractiveWebinar()->shouldBeCalled()->willReturn(false);
         $this->happening->isVideoWebinar()->shouldBeCalled()->willReturn(false);
         $this->happening->getLiveUrl()->shouldBeCalled()->willReturn(null);
+        $this->happening->isSidebarAllowed()->willReturn(true);
 
         $form = $this->prophesize(Form::class);
         $update = new Update($this->happening->reveal());